requirements.txt
Django==2.2.1
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
djangorestframework==3.7.3
psycopg2==2.7.3.2
requests-2.22.0